pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "My first AI agent"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "crewai>=0.159.0",
    "firecrawl-py>=2.16.3",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
